requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
confusable-homoglyphs==2.0.2
cryptography==2.1.2
Django==1.11.5
django-annoying==0.10.3
django-autoslug==1.9.3
django-bower==5.2.0
django-countries==5.0
django-crispy-forms==1.6.1
django-extensions==1.9.6
django-haystack==2.6.1
django-jinja==2.4.0
django-machina==0.5.6
django-mptt==0.8.7
django-registration==2.3
django-scheduler==0.8.6.1
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
icalendar==3.11.7
idna==2.6
Jinja2==2.9.6
markdown2==2.3.4
MarkupSafe==1.0
mercurial==4.3.3
ndg-httpsclient==0.4.3
olefile==0.44
Pillow==4.3.0
pusher==1.7.2
pyasn1==0.3.7
pycparser==2.18
pyOpenSSL==17.3.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
uWSGI==2.0.15
Whoosh==2.7.4
